+++ src/codec.ts
@@ -1,8 +1,8 @@
-import {List, Map} from "immutable";
+import {List, Map, Record} from "immutable";
 
 export interface Codec<T,U> {
-  encode(input: T): U;
-  decode(input: U): T;
+  encode(input: T, context?: any): U;
+  decode(input: U, context?: any): T;
 }
 
 export function isCodec(obj: any): obj is Codec<any,any> {
@@ -13,50 +13,60 @@ export function isCodec(obj: any): obj is Codec<any,any> {
 
 export function invert<T1,T2>(codec: Codec<T1,T2>): Codec<T2,T1> {
   return {
-    encode: (input) => codec.decode(input),
-    decode: (input) => codec.encode(input)
+    encode: (input, context) => codec.decode(input, context),
+    decode: (input, context) => codec.encode(input, context)
   };
 }
 
 export function compose<T1,T2,T3>(codec_a: Codec<T1,T2>, codec_b: Codec<T2,T3>): Codec<T1,T3> {
   return {
-    encode: (input) => codec_b.encode(codec_a.encode(input)),
-    decode: (input) => codec_a.decode(codec_b.decode(input))
+    encode: (input, context) => codec_b.encode(
+      codec_a.encode(input, context),
+      context
+    ),
+    decode: (input, context) => codec_a.decode(
+        codec_b.decode(input, context),
+        context
+    )
   }
 }
+
 
 export function identity<T>(): Codec<T,T> {
   return {encode: (input) => input, decode: (input) => input}
 };
 
-export function nullable<T,U>(codec: Codec<T,U>): Codec<T | null, U | null> {
+/**
+ * Uses the value stored in the context as identifier
+ * in order to supply a codecs value.
+ *
+ * On encode, the input's value is written to the context identifier,
+ * mutating the context.
+ *
+ * On decode, a value is read from the context.
+ * If the contextual value is a function, it is called with no arguments
+ * in order to generate a value for the field.
+ *
+ * @param {string} identifier
+ * @returns {Codec<T, undefined>}
+ */
+export function contextValue<T>(identifier: string): Codec<T, undefined> {
   return {
-    encode: (input) => input != null ? codec.encode(input) : null,
-    decode: (input) => input != null ? codec.decode(input) : null
-  }
-}
-
-export function array<T,U>(codec: Codec<T,U>): Codec<Array<T>,Array<U>> {
-  return {
-    encode: (input: Array<T>) => input.map((item) => codec.encode(item)),
-    decode: (input: Array<U>) => input.map((item) => codec.decode(item))
-  }
-}
-
-const listToArray: Codec<List<any>,Array<any>> = {
-  encode: (input) => input.toArray(),
-  decode: (input) => List(input)
-}
-
-export function list<T,U>(codec: Codec<T,U>): Codec<List<T>, Array<U>> {
-  return compose(listToArray, array(codec));
-}
-
-
-export function map<K, V1, V2>(codec: Codec<V1,V2>): Codec<Map<K,V1>,Map<K,V2>> {
-  return {
-    encode: (input) => input.map(codec.encode).toMap(),
-    decode: (input) => input.map(codec.decode).toMap()
+    encode: (input, context) => {
+      context[identifier] = input;
+      return undefined;
+    },
+    decode: (input, context) => {
+      let value = context[identifier];
+      if (value === undefined) {
+        throw `No value provided for identifier '${identifier}' in context.`;
+      }
+      if (typeof value === "function") {
+        return value() as T;
+      } else {
+        return value as T;
+      }
+    }
   }
 }