+++ src/json/interfaces.ts
@@ -1,1 +1,0 @@
-export type Json = null | boolean | number | string | any[] | {[k: string]: any};