+++ src/nullable.spec.ts
@@ -1,0 +1,24 @@
+import {nullable} from "./codec";
+
+describe('nullable', () => {
+  it('should convert a codec which does not accept \'null\' values into one which does', () => {
+    let nullSafeIdentity = {
+      encode: (value: any) => {
+        if (value === null)
+          throw 'Value cannot be null';
+        return value;
+      },
+      decode: (value: any) => {
+        if (value === null)
+          throw 'Value cannot be null';
+        return value;
+      }
+    };
+    expect(() => nullSafeIdentity.encode(null)).toThrow();
+    expect(() => nullSafeIdentity.decode(null)).toThrow();
+
+    let codec = nullable(nullSafeIdentity);
+    expect(codec.encode(null)).toBe(null);
+    expect(codec.decode(null)).toBe(null);
+  });
+});